Fix DDC mixing frequency to be float iso int
Uwe noticed that in ft037ar.prc the DDC synthesizer frequencies were all
"xx.00" whilst the channels should be tuned to "xx.50" MHz.

There was a bug in the DDC frequency programming code - it was truncating
the channel frequency to int because it was copy-pasted from the PFB case
where the integer channel frequency is used to look up the actual channel
number.