Added option to set postob/checkmk5 threshold
Previously vex2snap had a hard-coded minimum gap of 20s between previous scan end
and current scan start to decide whether to insert "postob + checkmk5" calls
in the .snp file.

Now made it into command line parameter - leading to being able to force
checkmk5/postob *always* by setting "--postob-duration=0" so that even for
back-to-back scans they get triggered.

This might not be a good idea but that is up to the user to decide, not the
code.